separated more parts of the source; TODO: create an async(io) request API
